crud operations implementation, history fixes
